+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file.
 
 - Added section for colors in utilities.
 - Added progress tracker component and documentation.
+- Added tooltip component.
+- Added media object component.
+- Added action list component.
 
 ### Changed